+++ docsassist/deployments.py
@@ -85,7 +85,6 @@ class DynamicSettings(BaseSettings):
 
 
 rag_deployment_env_name: str = "RAG_DEPLOYMENT_ID"
-grading_deployment_env_name: str = "GRADING_DEPLOYMENT_ID"
 app_env_name: str = "DATAROBOT_APPLICATION_ID"
 
 
@@ -94,14 +93,5 @@ class RAGDeployment(DynamicSettings):
         validation_alias=AliasChoices(
             "MLOPS_RUNTIME_PARAM_" + rag_deployment_env_name,
             rag_deployment_env_name,
-        )
-    )
-
-
-class GradingDeployment(DynamicSettings):
-    id: str = Field(
-        validation_alias=AliasChoices(
-            "MLOPS_RUNTIME_PARAM_" + grading_deployment_env_name,
-            grading_deployment_env_name,
         )
     )